Add feature flags

+++ src/components/Note.jsx
@@ -1,20 +1,31 @@
 import React from "react";
-import DeleteIcon from '@material-ui/icons/Delete';
+import DeleteIcon from "@material-ui/icons/Delete";
+import { useFeatureFlag } from "configcat-react";
 
 function Note(props) {
+  const {
+    value: isDeleteFeatureEnabled,
+    loading: isDeleteFeatureEnabledLoading,
+  } = useFeatureFlag("isDeleteFeatureEnabled", false);
 
-    function handleClick() {
-        props.deleteNote(props.id);
+  function handleClick() {
+    if (!isDeleteFeatureEnabled) {
+      return;
     }
-
-    return (
-        <div className="note">
-            <h1>{props.title}</h1>
-            <p>{props.content}</p>
-            <button onClick={handleClick}>
-                <DeleteIcon/>
-            </button>
-        </div>);
+    props.deleteNote(props.id);
+  }
+  
+  return (
+    <div className="note">
+      <h1>{props.title}</h1>
+      <p>{props.content}</p>
+      {!isDeleteFeatureEnabledLoading && (
+        <button onClick={handleClick} disabled={!isDeleteFeatureEnabled}>
+          <DeleteIcon />
+        </button>
+      )}
+    </div>
+  );
 }
 
 export default Note;
